==> analysis/02_non-model-analysis.Rmd <==
---
title: "Non-model analysis"
author: "Suparna Chaudhry and Andrew Heiss"
date: "`r format(Sys.time(), '%F')`"
output: 
  html_document:
    code_folding: hide
editor_options: 
  chunk_output_type: console
---

```{r load-libraries, message=FALSE}
knitr::opts_chunk$set(fig.retina = 2,
                      tidy.opts = list(width.cutoff = 120),  # For code
                      options(width = 120))  # For output

library(tidyverse)
library(stringr)
library(forcats)
library(scales)
library(patchwork)
library(countrycode)
library(sf)
library(here)

source(here("lib", "graphics.R"))
source(here("lib", "pandoc.R"))
source(here("lib", "bayes.R"))

my.seed <- 1234
set.seed(my.seed)
```

```{r load-data, cache=TRUE, message=FALSE}
df.donor <- readRDS(here("Data", "data_clean", "df_donor.rds"))
df.donor.us <- readRDS(here("Data", "data_clean", "df_donor_usaid.rds"))
df.country.aid <- readRDS(here("Data", "data_clean", "df_country_aid_no_imputation.rds"))

dcjw.questions.clean <- read_csv(here("Data", "data_manual", "dcjw_questions.csv"))
dcjw.responses.clean <- read_csv(here("Data", "data_manual", "dcjw_responses.csv"))

autocracies <- df.country.aid %>%
  group_by(cowcode) %>%
  summarise(polity.avg = mean(polity, na.rm = TRUE)) %>%
  mutate(autocracy = polity.avg < 6) %>%
  # These countries are missing polity scores
  mutate(autocracy = case_when(
    # Equatorial Guinea, Kuwait, Bahrain, UAE, Oman, Singapore
    .$cowcode %in% c(411, 690, 692, 696, 698, 830) ~ TRUE,
    # Serbia, Kosovo
    .$cowcode %in% c(340, 347) ~ FALSE,
    TRUE ~ .$autocracy
  ))
```

## Overall data summary

```{r data-summary}
num.countries <- df.country.aid %>% distinct(cowcode) %>% nrow()
num.years <- df.country.aid %>% distinct(year) %>% nrow()
year.first <- df.country.aid %>% distinct(year) %>% min()
year.last <- df.country.aid %>% distinct(year) %>% max()
```

Our data includes information about `r num.countries` countries across `r num.years` years (from `r year.first`–`r year.last`)


## Summary of variables in model

The values here are slightly different from what we had at ISA and MPSA (and our ISQ submission) because we're now using V-Dem 8.0 and AidData 3.1. 

```{r summary-vars-model, results="asis", message=FALSE}
coef.names <- read_csv(here("Data", "data_manual", "coef_names.csv"))

vars.to.summarize <- coef.names %>%
  filter(summarize == TRUE)

vars.summarized <- df.country.aid %>%
  select(one_of(vars.to.summarize$term)) %>%
  mutate(trade.pct.gdp = trade.pct.gdp / 100,  # This is scaled up for modeling
         total.oda = total.oda / 1000000) %>%
  gather(term, value) %>%
  filter(!is.na(value)) %>% 
  group_by(term) %>%
  summarize(N = n(),
            Mean = mean(value),
            Median = median(value),
            `Std. Dev.` = sd(value),
            Min = min(value),
            Max = max(value)) %>% 
  left_join(vars.to.summarize, by = "term") %>%
  arrange(summary_order) %>%
  select(Variable = term_clean_table, Source = source, 
         Mean, `Std. Dev.`, Median, Min, Max, N) %>%
  as.data.frame()

caption <- "Summary of all variables included in models {#tbl:var-summary}"
var.summary <- pandoc.table.return(vars.summarized, keep.line.breaks = TRUE, 
                                   round = 2, big.mark = ",",
                                   justify = "llcccccc", caption = caption, style = "multiline")

cat(var.summary)
cat(var.summary, file = here("Output", "tbl-var-summary.md"))
```


## Aid stuff

### Overall OECD aid

```{r summarize-oecd-aid}
df.country.aid %>%
  summarise(total = dollar(sum(total.oda))) %>% 
  pull(total) -> total.oecd.aid
```

OECD members donated `r total.oecd.aid` between 1981 and 2013.

```{r plot-oecd-aid}
plot.aid <- df.country.aid %>%
  group_by(year) %>%
  summarise(total = sum(total.oda)) %>% 
  mutate(fake_facet_title = "ODA commitments (all OECD)")

fig.oecd.aid <- ggplot(plot.aid, aes(x = year, y = (total / 1000000000))) + 
  geom_line(size = 0.5) +
  labs(x = NULL, y = "Billions of USD") +
  scale_y_continuous(labels = dollar) +
  theme_donors() + 
  facet_wrap(~ fake_facet_title)
fig.oecd.aid
```

### Proportion of contentious vs. noncontentious aid

```{r summarize-contentious-aid, results="asis"}
df.donor %>%
  filter(year > 1980) %>%
  group_by(purpose.contentiousness) %>%
  summarise(total = sum(oda)) %>%
  mutate(perc = total / sum(total)) %>%
  mutate(total = dollar(total), perc = percent(perc)) %>% 
  pandoc.table()
```

```{r plot-contentious-aid}
plot.aid.contentiousness <- df.donor %>%
  filter(year > 1980) %>%
  group_by(year, purpose.contentiousness) %>%
  summarise(total = sum(oda)) %>%
  mutate(perc = total / sum(total)) %>%
  filter(purpose.contentiousness == "High") %>% 
  mutate(fake_facet_title = "Contentious aid (all OECD)")

fig.oecd.contention <- ggplot(plot.aid.contentiousness, aes(x = year, y = perc)) + 
  geom_line(size = 0.5) +
  scale_y_continuous(labels = percent_format(accuracy = 1)) + 
  labs(x = NULL, y = "Percent") +
  theme_donors() +
  facet_wrap(~ fake_facet_title)
fig.oecd.contention
```

### USAID aid

```{r summarize-us-aid}
df.country.aid %>%
  summarise(total = dollar(sum(oda.us))) %>% 
  pull(total) -> total.us.aid

df.country.aid %>%
  filter(year > 1999) %>%
  summarise(total = dollar(sum(oda.us))) %>% 
  pull(total) -> total.us.aid.post.2000
```

The US donated `r total.us.aid` between 1981 and 2013 and `r total.us.aid.post.2000` between 2000 and 2013.

```{r plot-us-aid}
plot.aid.us <- df.country.aid %>%
  group_by(year) %>%
  summarise(total = sum(oda.us)) %>% 
  mutate(fake_facet_title = "ODA commitments (USAID only)")

fig.us.aid <- ggplot(plot.aid.us, aes(x = year, y = (total / 1000000000))) + 
  geom_line(size = 0.5) +
  expand_limits(y = 0) +
  labs(x = NULL, y = "Billions of USD") +
  scale_y_continuous(labels = dollar) +
  theme_donors() +
  facet_wrap(~ fake_facet_title)
fig.us.aid
```

### Proportion of US aid to types of NGOs

Total amounts over time:

```{r summarize-aid-channels, results="asis"}
usaid.total <- df.country.aid %>% summarise(total = sum(oda.us)) %>% pull(total)

df.country.aid %>%
  gather(channel, total.oda.us, c(oda.us.ngo.dom, oda.us.ngo.int, oda.us.ngo.us)) %>%
  group_by(channel) %>%
  summarise(total = sum(total.oda.us)) %>%
  mutate(perc = total / usaid.total) %>%
  mutate(total = dollar(total), perc = percent(perc)) %>% 
  pandoc.table()
```

The US clearly favors US-based NGOs or international NGOs over domestic NGOs.

```{r plot-aid-channels}
usaid.total.yearly <- df.country.aid %>%
  group_by(year) %>%
  summarise(annual.total = sum(oda.us)) %>% 
  mutate(fake_facet_title = "USAID ODA channeled through NGOs")

channels.nice <- tribble(
  ~channel,         ~channel.clean,
  "oda.us.ngo.dom", "Domestic NGOs",
  "oda.us.ngo.int", "International NGOs",
  "oda.us.ngo.us",  "US-based NGOs"
)

plot.usaid.channel <- df.country.aid %>%
  gather(channel, total.oda.us, c(oda.us.ngo.dom, oda.us.ngo.int, oda.us.ngo.us)) %>%
  group_by(year, channel) %>%
  summarise(total = sum(total.oda.us)) %>%
  left_join(usaid.total.yearly, by = "year") %>%
  mutate(perc = total / annual.total) %>%
  left_join(channels.nice, by = "channel")

fig.usaid.channel <- ggplot(plot.usaid.channel, 
                            aes(x = year, y = perc, colour = channel.clean)) + 
  geom_line(size = 0.5) +
  scale_y_continuous(labels = percent_format(accuracy = 1)) + 
  scale_colour_manual(values = channel.colors) +
  labs(x = NULL, y = "Percent") +
  guides(colour = guide_legend(title = NULL, reverse = TRUE, nrow = 2)) +
  theme_donors() + 
  facet_wrap(~ fake_facet_title)
fig.usaid.channel
```

USAID data categorizes all aid as government-channeled before 2000 because of some quirk in the data.

```{r plot-all-channels}
plot.usaid.channels.all <- df.donor.us %>%
  group_by(year, channel_subcategory_name) %>%
  summarise(total = sum(oda.us.2011)) %>%
  mutate(perc = total / sum(total)) %>%
  mutate(channel = ifelse(str_detect(channel_subcategory_name, "NGO"), "NGO", "Other"))

ggplot(plot.usaid.channels.all, 
       aes(x = year, y = perc, colour = channel_subcategory_name)) + 
  geom_line(size = 0.5) +
  scale_y_continuous(labels = percent_format(accuracy = 1)) + 
  labs(x = NULL, y = "Percent of US aid given through different channels") +
  guides(colour = guide_legend(title = NULL)) +
  theme_donors()
```

So we just look at aid after 2000.

```{r plot-channels-post-2000}
fig.usaid.channel.trimmed <- fig.usaid.channel +
  coord_cartesian(xlim = c(2000, 2013))
fig.usaid.channel.trimmed
```

### All DV figures combined

```{r plot-all-dvs, fig.width=5.5, fig.height=3.75}
fig.dvs <- (fig.oecd.aid + fig.oecd.contention) / 
  (fig.us.aid + fig.usaid.channel.trimmed) &
  theme(legend.text = element_text(size = rel(0.6)),
        axis.title.y = element_text(margin = margin(r = 3)),
        legend.box.margin = margin(t = -0.5, unit = "lines"))

fig.dvs
fig.save.cairo(fig.dvs, filename = "fig-dvs", 
               width = 5.5, height = 3.75)
```


## Legal restrictions on NGOs

### DCJW indexes

```{r dcjw-index-table, results="asis", message=FALSE}
dcjw.indexes <- read_csv(here("Data", "data_manual", "dcjw_index.csv")) %>%
  mutate(Laws = str_replace_all(Laws, "\n", "\n\n"))

caption <- "Description of indexes of NGO barriers {#tbl:ngo-barriers-index}"
dcjw.index.table <- pandoc.table.return(dcjw.indexes, keep.line.breaks = TRUE, 
                                        style = "grid", justify = "lll", caption = caption)

cat(dcjw.index.table)
cat(dcjw.index.table, file = here("Output", "tbl-ngo-barriers-index.md"))
```

### NGO barriers over time

```{r advocacy-laws, message=FALSE, fig.width=5.5, fig.height=4.5}
dcjw.questions <- read_csv(here("Data", "data_manual", "dcjw_questions.csv")) %>% 
  filter(!ignore_in_index) %>% 
  select(barrier_group = barrier_display, barrier = question_clean, 
         barrier_display = question_display) %>% 
  mutate(barrier_group = paste0("Barriers to ", str_to_lower(barrier_group)))

df.barriers <- df.country.aid %>% 
  group_by(cowcode, year) %>% 
  summarize_at(vars(one_of(dcjw.questions$barrier)), funs(. > 0)) %>% 
  group_by(year) %>% 
  summarize_at(vars(-cowcode, -year), funs(sum(.))) %>% 
  gather(barrier, value, -year) %>% 
  left_join(dcjw.questions, by = "barrier") %>% 
  mutate(barrier_display = str_replace(barrier_display, "XXX", "\n")) %>% 
  arrange(desc(value)) %>% 
  mutate(barrier_display = fct_inorder(barrier_display, ordered = TRUE))

dcjw_entry_plot <- ggplot(filter(df.barriers, 
                                 barrier_group == "Barriers to entry"), 
                          aes(x = year, y = value, 
                              color = barrier_display,
                              linetype = barrier_display)) +
  geom_line(size = 0.5) +
  expand_limits(y = c(0, 90)) +
  scale_y_continuous(sec.axis = sec_axis(~ . / num.countries,
                                         labels = percent_format(accuracy = 1)),
                     expand = c(0, 0)) +
  scale_colour_manual(values = c("black", "grey80", "grey50"), name = NULL) +
  scale_linetype_manual(values = c("solid", "solid", "21"), name = NULL) +
  guides(color = guide_legend(nrow = 2)) +
  labs(x = NULL, y = "Number of countries") +
  theme_donors() + 
  theme(legend.justification = "left") +
  facet_wrap(~ barrier_group)

dcjw_funding_plot <- ggplot(filter(df.barriers, 
                                   barrier_group == "Barriers to funding"), 
                            aes(x = year, y = value, 
                                color = barrier_display,
                                linetype = barrier_display)) +
  geom_line(size = 0.5) +
  expand_limits(y = c(0, 40)) +
  scale_y_continuous(sec.axis = sec_axis(~ . / num.countries,
                                         labels = percent_format(accuracy = 1)),
                     expand = c(0, 0)) +
  scale_colour_manual(values = c("black", "grey80", "grey50", "black", "grey80"), name = NULL) +
  scale_linetype_manual(values = c("solid", "solid", "solid", "21", "21"), name = NULL) +
  guides(color = guide_legend(nrow = 3),
         linetype = guide_legend(nrow = 3)) +
  labs(x = NULL, y = "Number of countries") +
  theme_donors() + 
  theme(legend.justification = "left") +
  facet_wrap(~ barrier_group)

dcjw_advocacy_plot <- ggplot(filter(df.barriers, 
                                    barrier_group == "Barriers to advocacy"), 
                            aes(x = year, y = value, 
                                color = barrier_display)) +
  geom_line(size = 0.5) +
  expand_limits(y = c(0, 40)) +
  scale_y_continuous(sec.axis = sec_axis(~ . / num.countries,
                                         labels = percent_format(accuracy = 1)),
                     expand = c(0, 0)) +
  scale_colour_manual(values = c("black", "grey80"), name = NULL) +
  guides(color = guide_legend(nrow = 1)) +
  labs(x = NULL, y = "Number of countries") +
  theme_donors() + 
  theme(legend.justification = "left") +
  facet_wrap(~ barrier_group)

df.csre.plot <- df.country.aid %>%
  left_join(autocracies, by = "cowcode") %>%
  group_by(year, autocracy) %>%
  nest() %>% 
  mutate(cis = data %>% map(~ mean_cl_normal(.$csre))) %>% 
  unnest(cis) %>% 
  mutate(fake_facet_title = "Civil society regulatory environment",
         autocracy = factor(autocracy, 
                            labels = c("Democracy (Polity IV ≥ 6)",
                                       "Non-democracy (Polity IV < 6)"), 
                            ordered = TRUE))

fig.csre <- ggplot(df.csre.plot, aes(x = year, y = y)) +
  geom_ribbon(aes(ymin = ymin, ymax = ymax, fill = autocracy), alpha = 0.2) +
  geom_line(aes(color = autocracy), size = 0.5) +
  annotate(geom = "text", x = 2013, y = -2.2, hjust = "right", size = 1.8,
           label = "Larger values = more open civil society") +
  scale_colour_manual(values = c("black", "grey75"), name = NULL) +
  scale_fill_manual(values = c("black", "grey75"), name = NULL) +
  scale_linetype_manual(values = c("solid", "solid", "21")) +
  labs(y = "Average CSRE", x = NULL) +
  theme_donors() +
  theme(legend.justification = "left") +
  facet_wrap(~ fake_facet_title)

barriers_summary <- 
  ((dcjw_entry_plot + dcjw_funding_plot) / 
     (dcjw_advocacy_plot + fig.csre)) &
  theme(legend.text = element_text(size = rel(0.6)),
        axis.title.y = element_text(margin = margin(r = 3)),
        legend.box.margin = margin(t = -0.5, unit = "lines"))

barriers_summary
fig.save.cairo(barriers_summary, filename = "fig-barriers-summary",
               width = 5.5, height = 4.5)
```

### Compulsory vs. burdensome registration

Laws requiring NGO registration aren't necessarily a sign of oppression—even the US requires that nonprofits that earn above a certain threshold register as 501(c)(3) organizations. Though the figure below shows that compulsory regulation have increased over time, actual restriction has occurred too. Burdensome registration is not just another standard layer of bureaucracy.

```{r compulsory-vs-burdensome, fig.width=4.5, fig.height=2}
df.regulation <- df.country.aid %>%
  left_join(autocracies, by = "cowcode") %>%
  group_by(year, autocracy) %>%
  summarise(`Registration required` = sum(ngo_register) / n(),
            `Registration burdensome` = sum(ngo_register_burden) / n()) %>%
  gather(type.of.law, value, -year, -autocracy) %>%
  mutate(autocracy = 
           factor(autocracy, levels = c(TRUE, FALSE),
                  labels = c("Non-democracies", "Democracies")))

fig.regulation.burden <- ggplot(df.regulation, 
                                aes(x = year, y = value, colour = type.of.law)) +
  geom_line(size = 0.5) +
  scale_y_continuous(labels = percent_format(accuracy = 1)) +
  scale_x_continuous(expand = c(0, 0)) +
  coord_cartesian(ylim = c(0, 0.7), xlim = c(1980, 2015)) +
  scale_colour_manual(values = burden.colors) +
  guides(colour = guide_legend(title = NULL)) +
  labs(x = NULL, y = "Proportion of countries\nwith regulation") +
  theme_donors() +
  facet_wrap(~ autocracy)

fig.regulation.burden
fig.save.cairo(fig.regulation.burden, filename = "fig-regulation-burden",
               width = 4.5, height = 2)
```


## Aid

### Aid over time, by donor type

```{r aid-by-donor}
aid.donor.type.time <- df.donor %>%
  group_by(year, donor.type.collapsed) %>%
  summarise(total.aid = sum(oda, na.rm = TRUE))

ggplot(aid.donor.type.time, aes(x = year, y = total.aid / 1000000000,
                                colour = donor.type.collapsed)) +
  geom_line(size = 0.5) +
  labs(x = NULL, y = "Billions of USD",
       caption = "Source: OECD and AidData. 2011 dollars.") +
  guides(colour = guide_legend(title = NULL)) +
  scale_y_continuous(labels = dollar) + 
  theme_donors()
```

### Aid over time, by contentiousness

```{r aid-by-contention}
aid.contention.time <- df.donor %>%
  group_by(year, purpose.contentiousness) %>%
  summarise(total.aid = sum(oda, na.rm = TRUE))

ggplot(aid.contention.time, aes(x = year, y = total.aid / 1000000000,
                                colour = purpose.contentiousness)) +
  geom_line(size = 0.5) +
  labs(x = NULL, y = "Billions of USD",
       caption = "Source: OECD and AidData. 2011 dollars.") +
  guides(colour = guide_legend(title = NULL)) +
  scale_y_continuous(labels = dollar) + 
  theme_donors()
```

## Restrictions and aid {.tabset .tabset-fade .tabset-pills}

```{r restrictions-aid-correlations}
inv.logit <- function(f, a) {
  a <- (1 - 2 * a)
  (a * (1 + exp(f)) + (exp(f) - 1)) / (2 * a * (1 + exp(f)))
}

dvs.clean.names <- tribble(
  ~key, ~key.clean,
  "barriers.total", "All barriers",
  "advocacy", "Barriers to advocacy",
  "entry", "Barriers to entry",
  "funding", "Barriers to funding"
)

ivs.clean.names <- tribble(
  ~variable, ~variable.clean, ~hypothesis,
  "total.oda_log_next_year", "Total ODA", "H1",
  "prop.contentious_logit_next_year", "Contentious aid", "H2",
  "prop.ngo.dom_logit_next_year", "Aid to domestic NGOs", "H3",
  "prop.ngo.foreign_logit_next_year", "Aid to foreign NGOs", "H3"
)
```

### Restrictions and ODA (H~1~)

```{r restrictions-aid-h1, warning=FALSE, fig.width=8, fig.height=4.5}
df.plot.barriers.oda <- df.country.aid %>% 
  select(year, cowcode, country.name, total.oda_log_next_year,
         one_of(dvs.clean.names$key)) %>% 
  gather(key, value, one_of(dvs.clean.names$key)) %>% 
  filter(!is.na(total.oda_log_next_year), !is.na(value)) %>% 
  mutate(total.oda.transformed = expm1(total.oda_log_next_year)) %>% 
  left_join(dvs.clean.names, by = "key") %>% 
  mutate(key.clean = fct_inorder(key.clean, ordered = TRUE))

ggplot(df.plot.barriers.oda, 
       aes(x = value, y = total.oda.transformed, color = key.clean)) +
  geom_point(alpha = 0.5) +
  stat_smooth(method = "lm") +
  stat_smooth(data = filter(df.plot.barriers.oda, 
                            total.oda.transformed > 10000000000), 
              method = "lm", linetype = "21") +
  scale_y_continuous(labels = dollar) +
  guides(color = FALSE) +
  labs(x = "Number of legal barriers", y = "Total ODA in next year",
       title = "Total ODA in next year",
       subtitle = "Dotted lines show trends when omitting observations\nwith less than $10,000,000,000 in ODA") +
  theme_donors() +
  theme(strip.text.x = element_text(margin = margin(t = 1, b = 1))) +
  facet_wrap(~ key.clean, scales = "free_x", nrow = 2)
```

### Restrictions and contentiousness (H~2~)

```{r restrictions-aid-h2, warning=FALSE, fig.width=8, fig.height=4.5}
df.plot.barriers.contention <- df.country.aid %>% 
  select(year, cowcode, country.name, prop.contentious_logit_next_year,
         one_of(dvs.clean.names$key)) %>% 
  gather(key, value, one_of(dvs.clean.names$key)) %>% 
  filter(!is.na(prop.contentious_logit_next_year), !is.na(value)) %>% 
  mutate(prop.contentious.transformed = 
           inv.logit(prop.contentious_logit_next_year, a = 0.001)) %>% 
  left_join(dvs.clean.names, by = "key") %>% 
  mutate(key.clean = fct_inorder(key.clean, ordered = TRUE))

ggplot(df.plot.barriers.contention, 
       aes(x = value, y = prop.contentious.transformed, color = key.clean)) +
  geom_point(alpha = 0.5) +
  stat_smooth(method = "lm") +
  stat_smooth(data = filter(df.plot.barriers.contention, 
                            prop.contentious.transformed > 0.05), 
              method = "lm", linetype = "21") +
  scale_y_continuous(labels = percent) +
  guides(color = FALSE) +
  labs(x = "Number of legal barriers", 
       y = "Proportion of contentious aid in next year",
       title = "Proportion of contentious aid in next year",
       subtitle = "Dotted lines show trends when omitting observations\nwith less than 5% contentious aid") +
  theme_donors() +
  theme(strip.text.x = element_text(margin = margin(t = 1, b = 1))) +
  facet_wrap(~ key.clean, scales = "free_x", nrow = 2)
```

### Restrictions and NGOs (H~3~)

```{r restrictions-aid-h3, warning=FALSE, fig.width=8, fig.height=4.5}
df.plot.barriers.ngos <- df.country.aid %>% 
  select(year, cowcode, country.name, 
         prop.ngo.dom_logit_next_year, prop.ngo.foreign_logit_next_year,
         one_of(dvs.clean.names$key)) %>% 
  gather(barrier, value, one_of(dvs.clean.names$key)) %>% 
  gather(variable, prop.ngo, prop.ngo.dom_logit_next_year,
         prop.ngo.foreign_logit_next_year) %>% 
  filter(!is.na(prop.ngo)) %>% 
  mutate(prop.ngo.transformed = inv.logit(prop.ngo, a = 0.001)) %>% 
  left_join(dvs.clean.names, by = c("barrier" = "key")) %>% 
  left_join(ivs.clean.names, by = "variable") %>% 
  mutate(key.clean = fct_inorder(key.clean, ordered = TRUE))

ggplot(df.plot.barriers.ngos, 
       aes(x = value, y = prop.ngo.transformed, color = key.clean)) +
  geom_point(alpha = 0.5) +
  stat_smooth(method = "lm") +
  stat_smooth(data = filter(df.plot.barriers.ngos, 
                            prop.ngo.transformed > 0.05), 
              method = "lm", linetype = "21") +
  scale_y_continuous(labels = percent) +
  guides(color = FALSE) +
  labs(x = "Number of legal barriers", 
       y = "Proportion of aid to NGOs in next year",
       title = "Proportion of aid channeled to types of NGOs in next year",
       subtitle = "Dotted lines show trends when omitting observations\nwith less than 5% aid to NGOs") +
  coord_cartesian(ylim = c(0, 1)) +
  theme_donors() +
  theme(strip.text.x = element_text(margin = margin(t = 1, b = 1))) +
  facet_wrap(~ variable.clean + key.clean, scales = "free_x", ncol = 4)
```

### CSRE and all DVs (all hypotheses)

```{r csre-aid-allhs, warning=FALSE, fig.width=8, fig.height=4.5}
df.plot.csre.ngos <- df.country.aid %>% 
  select(year, cowcode, country.name, 
         prop.contentious_logit_next_year, total.oda_log_next_year,
         prop.ngo.dom_logit_next_year, prop.ngo.foreign_logit_next_year,
         csre) %>% 
  gather(variable, value, -c(year, cowcode, country.name, csre)) %>% 
  filter(!is.na(value)) %>% 
  left_join(ivs.clean.names, by = "variable") %>% 
  mutate(hypothesis.clean = paste0(hypothesis, ": ", variable.clean)) %>% 
  arrange(hypothesis.clean) %>% 
  mutate(hypothesis.clean = fct_inorder(hypothesis.clean, ordered = TRUE)) %>% 
  mutate(value.transformed = case_when(
    .$hypothesis == "H1" ~ expm1(.$value),
    .$hypothesis == "H2" ~ inv.logit(.$value, a = 0.001),
    .$hypothesis == "H3" ~ inv.logit(.$value, a = 0.001),
  ))

ggplot(df.plot.csre.ngos, 
       aes(x = csre, y = value.transformed, color = hypothesis)) +
  geom_point(alpha = 0.25) +
  scale_color_viridis_d(option = "plasma", end = 0.9) +
  guides(color = FALSE) +
  labs(x = "Civil society regulatory environment", 
       y = "Variable value in next year",
       title = "Civil society regulatory environment") +
  theme_donors() +
  facet_wrap(~ hypothesis.clean, scales = "free_y")
```


## CIVICUS restrictions

```{r civicus-data, warning=FALSE, message=FALSE}
civicus <- read_csv(here("data", "data_raw", "Civicus", "civicus_monitor_2017.csv"),
                    na = "Null") %>%
  mutate(Population = as.double(Population),  # Integers can't handle world population
         Rating = factor(Rating, levels = c("Open", "Narrowed", "Obstructed", 
                                            "Repressed", "Closed"), 
                         ordered = TRUE),
         iso3 = countrycode(Country, "country.name", "iso3c"))

# Robinson projection
projection = 54030

world_shapes <- st_read(file.path("data", "data_raw", "ne_110m_admin_0_countries",
                                  "ne_110m_admin_0_countries.shp"),
                        quiet = TRUE) %>% 
  filter(ISO_A3 != "ATA")
```

```{r civicus-numbers, results="asis"}
civicus %>% count(Rating) %>% pandoc.table()
```

```{r civicus-map, fig.width=5.5, fig.height=3}
map_with_civicus <- world_shapes %>% 
  # Fix some Natural Earth ISO weirdness
  mutate(ISO_A3 = ifelse(ISO_A3 == "-99", as.character(ISO_A3_EH), as.character(ISO_A3))) %>% 
  mutate(ISO_A3 = case_when(
    .$ISO_A3 == "GRL" ~ "DNK",
    .$NAME == "Norway" ~ "NOR",
    TRUE ~ ISO_A3
  )) %>% 
  left_join(civicus, by = c("ISO_A3" = "iso3"))

plot_civicus_map <- ggplot() +
  geom_sf(data = map_with_civicus, aes(fill = Rating), size = 0.15, color = "black") +
  coord_sf(crs = st_crs(projection), datum = NA) +
  scale_fill_manual(values = c("grey90", "grey70", "grey45",
                               "grey20", "black"),
                    na.translate = FALSE, name = "Civic space") +
  theme_donors_map() + theme(legend.key.size = unit(0.7, "lines"))

plot_civicus_map 
fig.save.cairo(plot_civicus_map, filename = "fig-civicus-map",
               width = 5.5, height = 3)
```


## List of countries included in models

```{r list-countries, results="asis"}
matrix_from_vector <- function(x, ncol) {
  n_balanced <- ceiling(length(x) / ncol) * ncol
  matrix(c(x, rep(NA, n_balanced - length(x))), ncol = ncol)
}

all_countries <- df.country.aid %>% 
  distinct(country.name) %>% 
  arrange(country.name) %>% 
  pull(country.name) 

caption <- paste0("All countries included in models (N = ", 
                  length(all_countries),
                  ") {#tbl:countries}")

ncol_countries <- 4

tbl_countries <- all_countries %>% 
  matrix_from_vector(ncol = ncol_countries) %>% 
  pandoc.table.return(justify = paste0(rep("l", ncol_countries), collapse = ""), 
                      caption = caption, missing = "")

cat(tbl_countries)
cat(tbl_countries, file = here("Output", "tbl-countries.md"))
```
